Add support for responder identity authentication
During this phase a requester retrieves digests of up to 8 responder certificate
chains, not including the root cert. The requester than goes ahead and requests
the chains. These cert chains are used in later parts of the protocol.

While the protocol allows up to 8 certificate slots, we allow the user to limit
this number to lower memory usage. This is necessary for ergonomic usage in
no_std environments, that consists of stack allocating arrays in protocol code.
The protocol also allows cert chains to be up to 65536 bytes long. To help limit
memory usage in constrained environments, and the likelihood that the chains are
going to be much smaller in practice, we allow the user to limit the max size of
these chains.

This commit also contains the first usage of crypto code in the form of Digests.
A new `Digest` trait was created and backed by Ring.

Integration tests were also cleaned up to make them easier to understand.

A few shortcuts were taken that are not crucial to the work being done right now, and will be fixed in the future:

 * Configuration is hardcoded in config.rs
 * The requester doesn't yet verify that retrieved cert chains match retrieved digests.
 * The code assumes that cert chains can always be retreived in one message.
 * Ring usage is not behind a feature flag