Add Challenge Authentication support
PKI related functionality based on
[webpki](https://github.com/briansmith/webpki) is used to validate received
certificates and verify signed messages in the requester. An `EndEntityCert`
trait is used to abstract the underlying PKI mechanism. Although, currently,
there is only a single concrete type (`WebpkiEndEntityCert`), it is anticipated
that there will be more. Due to the lack of Generic Associated Types (GATs), in
stable rust, the strategy of selecting a concrete PKI type via the `Config`
type, as used with `Digest`, was made damn near impossible. This is primarily
due to the lifetime parameter of `EndEntityCert`. Therefore construction of the
concrete type is done via a function, `new_end_entity_cert`, that returns the
proper concrete type. A feature will be added to allow users to select the
proper pki type at compile time. As more crypto types will probably borrow data,
it is likely that the `Config` type will be removed altogether in the future and
`Digest` will use the same pattern here. Note that, this is not really a
problem, as in many cases only a limited set of implementations will
compile.

While `webpki` supports RSA, it only does so via `alloc`, as it utilizes the
ring implementation. In order to make all code in this crate `no_std`, only 2
ECDSA algorithms are supported.

Due to the fact that all of the crypto in this crate is going to be configured
via feature flags, the set of supported algorithms should be generated at build
time for the user, or at least made available in an easy to use manner. The user
may choose to further restrict algorithms beyond what the underlying platform is
capable of and we should allow this. This is just a note for the future, as it
is not going to be done in this PR.

A new `test-utils` crate was added to this repository in order to facilitate
sharing code between integration and unit tests. All the certificate generation
test code based around rcgen was moved in there.