Enhance the ergonomics of the Responder (#27)
Prior to this commit, all the internal states of the Responder were exposed to
the user, in a typestate style API. While safe, this is a bit unwieldy and
requires every user to duplicate the flow of the state machine. This commit
simplifies the user API dramatically such that users now only have to make a
single API call, `Responder::handle_msg`, when a message is received over the
transport.

In order to make this code compile, and avoid borrow check issues, the internal
states once again return the amount of data written rather than a slice of the
output buffer. However, a slice is still presented in the user API, maintaining
the ergonomics of prior versions.

Since we now have an `AllStates` wrapper, the individual `Transition` enums for
each state also are gone, reducing the amount of state specific code necessary.

This change also provides conversion between the `ResponderError` and the spdm
Error message in `msgs::Error`. If an error is returned, the spdm error message
is written to the output buffer similar to any successful response.