Improve ergonomics of the Requester (#32)
All internal, capability driven request response messages of the SPDM protocol
have a unique `State` as part of a typestate pattern. The prior version of this
code required the user to manage the transition between all of these requester
states manually. Furthermore, each state could require different arguments
specific to the message sequence being exchanged. This required the user of this
library to have internal knowledge of the SPDM protocol, and also resulted in
much duplication of code by different users. The upshot of this strategy is that
the Rust type system made it impossible to send an invalid request at a given
state of the protocol, and the handling code automatically checks that the
expected response is received. The goal of this change then is to maintain this
safety while reducing unnecessary user code duplication and complexity.

The key insight behind this change is to realize there are 2 distinct phases to
the SPDM protocol:
 * Secure session initialization
 * Application level messages inside a secure session

Secure session initialization can have its state transitions automated using
user configuration detailing capabilities and algorithms. Application level
messaging, including on demand measurement requests, must be driven by the user.

We bundle the initialization into the `RequesterInit` state. Users interact by
repeatedly calling `next_request` to serialize a request to send, and then
handling the response in the `handle_msg` method. Note that we still do not rely
on any given transport implementation or even trait here, and thus maintain the
ability to use SPDM in both no_std and async contexts.

Once session initialization is complete, `handle_msg` returns `Ok(true)`
instructing the user to transition to the `RequesterSession` state. Note that
this state is not fully implemented, and neither is the actual ability to setup
a secure session. This code just lays out the patterns needed and transitions
the existing code base.  One could ask: "Why not transition directly to the
RequesterSession state automatically, following a more typical typestate
pattern?". The answer is mainly ergonomics again. There are many initialization
messages, and it's somewhat easier to check a bool than issue a match after each
`handle_msg` call.This is easy to change later, if we decide this isn't the
right decision though.