build: Disable deafult features on rand dependency.
The rand crate follows the convention of having an 'std' feature that
enables stuff that depends on 'std'. This feature is enabled by default.
Disabling the default features for this crate allows us to build for
targets that aren't supported by the std lib (thumbv8m.main-none-eabihf
in my case).