Support crypto providers that implement the RustCrypto `Digest` trait
This is the first commit to start standardizing on crypto providers that
implement the RustCrypto traits. Providers that implement the `Digest` trait are
now supported. We use ring-compat to allow `Digest` to be suppported by our ring
provider.

Since the SPDM protocol allows selection of a digest at runtime, we cannot
hardcode which digest will be chosen (if there is more than one supported). This
means that we must wrap all possible digests implementations in an enum, so the
proper one can be selected at runtime. Additionally, since the return values of
many `Digest` methods differ in size, we cannot directly map the `Digest` trait
to the enum and delegate calls 1:1. We therefore take a simpler approach, and
create our own wrapper trait, `Digests`, that includes only necessary
functionality like incremental update and convenience one off hashing. We
provide a macro, `impl_digests`, that can be used to wrap an Enum of a certain
shape, where each variant is a tuple type with one element, and where that
element implements the RustCrypto `Digest` trait (or equivalent API, as the
`Digest` trait is not currently constrained). See the comments above the
`Digests` trait and `impl_digests` macro for more details.

We use the `impl_digests` macro to implement a `ProvidedDigests` enum around
`ring` digests, which are themselves wrapped by `ring-compat`.

Additionally, now that we can configure a digest provider, we no longer require
the user to manually configure hash algorithms. We make available all that are
supported by the provider. We may wish to one day allow restriction of supported
algorithms, but that is not urgent. You can see where this change is reflected
in the `config_to_negotiated_algorithms` function in
`src/requester/algorithms.rs`.  When all crypto mechanisms properly use
providers we can get rid of their configuration and change the name of this
function.

The unused `Provider` trait was also removed.